+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [tanjun.clients.Client.set_human_only][] no longer registers the internal check multiple
+  times when called with [True][] multiple times.
+
 ### [2.11.1] - 2023-01-07
 ### Added
 - Support for specifying Guild Forum channels for the constraints of a channel command option.